split files / cleanup
